Extract OCR functions and add unit tests for them.
_init_ file was emptied and it's contents moved into main.py. That was needed for the new unit test structure.
